refactor wiki dashboard into modular pages
